libkrad: Add Message-Authenticator support
draft-ietf-radext-deprecating-radius-03 describes multiple mitigation
measures for recent vulnerabilities discovered in the RADIUS protocol,
especially the BlastRADIUS one (CVE-2024-3596). One of them is the
enforcement of the Message-Authenticator attribute.

Message-Authenticator was initially defined in RFC2869 (2000) as an
extension to improve authenticity and integrity verification.

However the BlastRADIUS vulnerability is now pushing vendors to
generalize the use of this attibute. FreeRADIUS is now enforcing
Message-Authenticator since versions 3.2.5 and 3.0.27. libkrad has to
support Message-Authenticator in order to remain compatible with these
implementations.